sstatesig: do_image_complete needs to run with pseudo disbale
sstate_report_unihash function from sstate.bbclass needs to
run with pseudo disbaled when running do_image_complete task.
This is because it have the deploy content with uid and gid
of the user that run bitbake on the build machine.

By default the creation of sstate is skiped in the task do_image_complete
in the image.bbclass using SSTATE_SKIP_CREATION:task-image-complete=1
so the issue is not triggered.

There are some situation when is useful to store the image artifacts on the
sstate cache for future reuse and speedup the build without need to pruduce
the same image again. These cases are especially useful in the case of the
initramfs images used by the kernel.

The issue can be reproduced in oe-core with:
 echo "SSTATE_SKIP_CREATION:task-image-complete = '0'" >> meta/recipes-core/images/core-image-minimal-initramfs.bb
 bitbake core-image-minimal-initramfs -C do_image_complete

Signed-off-by: Jose Quaresma <jose.quaresma@foundries.io>